requirements.txt
pygame
requests
Pillow
tkcalendar
PySide6
pytesseract
